Update 0.27.5 (Fixed simple session attributes not working on google assistant, improved stability of variables access in the backend)
The simple session attributes (forgotten after the close of the session with no ability to use a timeout) actually never worked on google assistant, yet gave the illusion to work when i was using a lot of staticness in the code. Now they are fully working, mainly thanks to setting the lifespanCount variable to 999.
I improved the stability of accessing variables, by defining them as properties, and if they are none (or not of the right instance), they will initiate themselves, and then return the object.
I also started to create the smart session attributes (a mix of session attributes and persistent attributes, they will be saved in the database, and will have a timeout value. For example, if the timeout is 60 seconds, if the user close the skill and re-open it before 60 seconds have passed, the session will resume from the smart session attributes, without needing to write any additionnal code. Otherwise, it will not use the smart session attributes, considering that the user is not in the same session. It will be extremly useful since the session of an user, might include multiple launchs of the skill, and it will also avoid frustrations of relaunching from start if the skill close (which happens all the time x)  ).