0.28 Smart session attributes and session and persistents attributes for Bixby
The smart session attributes are like session attributes, but are stored in a database, and can will be accessed if a timeout has not been reached. If you set a timeout of 60 seconds, and the user close your skill, and re-open it before 60 seconds have passed, the framework will consider that the user is in the same session and use the smart session attributes from the previous session. Otherwise, it will clear the smart session attributes.
For now it only work for smart session attributes, yet the idea is that if the timeout has not been reached, the user will be resumed where he was at in the previous session, potentially by saying a different message to him that could be set in each Request/StateHandler.
The smart session attributes are now used by default when using self.session_remember(), self.session_memorize(), etc. If you want to use the old session_attributes that are passed as a json response to Alexa/Google-Assistant, you can use the self.simple_session_memorize(), self.simple_session_remember(), etc.
Warning, Bixby will not work with the simple_session_attributes, it will only work with the smart one (i highly recommand to use the smart one, if you want to do as if they were classic attributes, set the timeout to 0 instead).