0.29 (Auto-resuming to intents of the previous session of the user (a timeout can be specified on the session))
Little update, that i found important, is the ability (without extra code) for an user to resume to its previous state in the skill.

On both the InoftRequestHandler and InoftStateHandler classes, a new function has been added, the
def handle_resume(self):
    self.say("i should do things")
    self.to_platform_dict()

Each time an handler is used for the user (a request or state handler), they will be saved in the smart session attributes (that has a timeout that can is customizable). For example if you have the default timeout of 60 seconds, if the user launch your skills and for example say "Yes" to your intro, he might enter a class called YesHandler
Then, if the user close your skill, but comeback before the 60 seconds have elapsed, the skill will remember that he was in the YesHandler class, but instead of calling the handle function, it will call the handle_resume function.
That's the place where you woud say """ Hey, we left each other in the middle of something, do you want to resume it where you were ? """

Bonus : since i now have the optic fiber internet connection of Paris, and that it works less well than ADSL (seriously), i added the ability when create the InoftSkill() to use the argument disable_database=bool, that do what it say (all the session and persistent attributes will stop working, without giving exceptions and accepting their fates). I added that, because if you are not connected to internet, even in development mode, if you launch the skill it will get stuck at connecting with the database.