Update 0.34 (unstable google assistant carousels)
New function :
self.google.show_browse_carousel_item(self, title: str,  url_to_open_on_click: str, description: str = None, image_url: str = None, image_accessibility_text: str = None, footer: str = None):

No need to create a carousel object and add items to it, just add your items with the same arguments as the first item (if you have put an image on the first item, you must put an image on all of them).
If you set only one item, it will be duplicated to have 2 items (the minimum required). And any items over the 10th item will not be included (yet not throw an exception).

I'm saying that this update is "unstable", because for this update, the framework is not checking if the feature is available on the user device. The carousel is available only on devices that support a web browser (like a phone, where it will work just fine). But a smart speaker do not support a web browser, and if you use the carousel and try to test it with a smart speaker (or in the emulator and you select a smart speaker), it will give you an error that the device is not supported.

I will make it so that the carousel will be used only if the platform of the user can support it.