Added requirements to README
My next step is to improve the section on deploying and configuring botocore. Because currently what i'm saying is wrong. Due to an issue, if you have not configured botocore to use your AWS account, and that you try to deploy your project, you will basicly get no help whatsoever.