Update 0.7 (First integration with BotPress, a visual tool to create bots)
I'm doing the integration with BotPress instead of another service like Voicify or Voiceflow because :
- It is open-source
- It's free
- It has more functionnalities
- It's using situationnal design (way superior to any type of mindmap)

So even if BotPress is in my opinion the best tool if you are serious about vocal assistant developement, its called BotPress and not VocalPress ! This means that even if it allows us at Inoft to create huge stories with an unlimited amount of interconnections, it can only create text bot for Facebook/Skype/Messenger and the classic one.

This is only the first version of integration with the Inoft Vocal Framework. Currently its even more a code generator from your BotPress files, than an integration. Yet, let me tell you, at Inoft we are currently working on the first story of a game around WWII, where in around 10-15 minutes of gameplay, there will be hundreds of interactions with the user. We would not be able to do it without using BotPress and the Inoft Vocal Framework Integration.

Tutorials on the framework in general, BotPress and the integration are coming soon ! :)
Robinson Labourdette