Update 0.72 (20x invocation performance improvement)
Yes, each invocation of a skill with the inoft vocal framework is around 20 times as fast as previously.
So for example, for a big game on an AWS Lambda with 256 mb of memory, which took around 1.2 seconds to execute and return a response, will now take 50ms (these are really the results i'm getting).
We achieve that by initializing the database client only on the first invocation on the app, and then on futur invocation, we re-use the existing database client (that is stored in a static variable and shared across invocations).
So it is normal if the first invocation of your app is a bit slow (~1-2 seconds) after updating your code, since its initializing the database client, yet after that, its will go fast.
No modification on any existing code is required, only the framework has change, not any of its api or cli.