Update 0.73 (New class type InoftCondition and new Facts app sample)
The InoftCondition is a class that will only have the can_handle function. It's useful for the frequent things that an user can say at multiple places. Like Yes, No, etc.
Example :
class YesCondition(InoftCondition):
    def can_handle(self) -> bool:
        return self.is_in_intent_names(["AMAZON.YesIntent", "YesIntent"])

class SayFactOnYesStateHandler(InoftStateHandler):
    def handle(self) -> dict:
        if YesCondition(self).can_handle():
            return GetFactRequestHandler(self).handle()
        else:
            self.say("I'm supposed to be here to give you facts. Yet if i want i can give you some infos about this application. What do you say ? Yes ?")
            return self.to_platform_dict()

The facts sample can be selected when creating a new project with the inoft new command.
I'm starting to make a basic tutorial (in english and french) on how to create a facts app with the inoft vocal framework, so i wanted to make sure that everything was working correctly, and it does ! :)