Update 0.81 (Improved BotPress code generation)
Fixed somes rare bugs with the code generation, and also implemented the handling of the "WaitForUserResponse" or not option. It makes the code generation way more modulable, since you can have the same response that will be created by multiples nodes.
I also do some smalls improvements on the generation of the class names.